Analytical Report
In this report, we present our work on developing a machine learning model for the classification of dry beans. The dataset used in this study contains 13611 instances of dry beans, which are classified into 7 classes based on their characteristics such as shape, size, and color. Our goal is to develop a model that accurately classifies dry beans into their respective classes.
Previous research by Murat Koklu, Ilker Ali Ozkan just implemented 4 models including DT, MLP, DT, KNN but in our research we worked on more models which are explained in this report.
The report is organized as follows: In Section 2, we explore the dataset and provide a critique of its limitations. In Section 3, we describe the methods and techniques used in our work and provide justifications for each method choice. In Section 4, we present the results of our experiments using appropriate metrics and plots. In Section 5, we discuss our findings, limitations of our work for real-world applications, and compare our results with previous work. In Section 6, we report the development and evaluation of an additional automatic detection model using a different machine learning algorithm and compare its performance with our primary model. Finally, in Section 7, we provide annotated code and classification output.